Split tests related to virtualfiles into multiple test files
